added some fonts, made button components

+++ src/routes/admin.js
@@ -10,25 +10,36 @@ client.on('error', (err) => {
 await client.connect();
 
 /** @type {import('@sveltejs/kit').RequestHandler} */
-export async function post( {params} ) {
-
-    console.log("" + params.pass);
-
-    /* if(await params.get('pass') == '123') {
-
-        
-
-        let code = await params.get('code');
-        switch(code) {
-            case "flush":
-                client.flushAll();
+export async function post( {request} ) {
+    //get data from post request
+    let data = await request.formData();
+    //check if password is entered
+    let pass = await data.get('pass');
+    if (pass === 'trey') {
+        let code = await data.get('code') ?? '';
+        switch ( code) {
+            case 'flush':
+                await client.flushAll();
+                return {
+                    body: 'list flushed'
+                };
                 break;
-            case default:
+            case 'listall':
+                let colors = await client.LRANGE('color', 0 ,-1);
+                return {
+                    body: colors
+                };
                 break;
-        } 
-    } */
-
-    return {
-        body : 'hello'
+            default:
+                return {
+                    body: 'code not recognized'
+                };
+                break;
+        }
+    } else {
+        return {
+            status: 400,
+            body: 'wrong password'
+        }
     }
 }
